+++ component.json
@@ -25,6 +25,7 @@
   "scripts": [
     "index.js",
     "lib/errors.js",
+    "lib/view.js",
     "lib/xregexp-all.js"
   ],
   "templates": [
